requirements.txt
python-docx
python-pptx
lxml
pandas
jupyterlab
ipykernel
openpyxl
pydantic 
tqdm
